refactor(processors): move processors to folders named with priority
