+++ src/pwa/stores/server.js
@@ -2,8 +2,8 @@
 import { flow, getEnv } from 'mobx-state-tree';
 import base from '../../shared/stores';
 import { home, singleWithLatest } from '../../shared/contexts';
-import processors from '../../shared/processors';
-import converters from '../../shared/converters';
+import mediumProcs from '../../shared/processors/medium';
+import lowProcs from '../../shared/processors/low';
 
 export default base.actions(self => ({
   fetchSelectedItem: flow(function*() {
@@ -34,7 +34,7 @@ export default base.actions(self => ({
     // Set first context.
     yield self.fetchSelectedItem();
 
-    processors.forEach(proc => self.h2r.addProcessor(proc, 'medium'));
-    converters.forEach(conv => self.h2r.addProcessor(conv, 'low'));
+    mediumProcs.forEach(proc => self.h2r.addProcessor(proc, 'medium'));
+    lowProcs.forEach(conv => self.h2r.addProcessor(conv, 'low'));
   }),
 }));
