+++ src/shared/processors/low/audio.js
@@ -1,4 +1,4 @@
-import LazyAudio from '../components/LazyAudio';
+import LazyAudio from '../../components/LazyAudio';
 
 export default {
   test: ({ component }) => component === 'audio',
