+++ src/shared/processors/low/image.js
@@ -1,8 +1,8 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import Link from '../../pwa/components/Link';
-import Image from '../components/Image';
-import { media } from '../contexts';
+import Link from '../../../pwa/components/Link';
+import Image from '../../components/Image';
+import { media } from '../../contexts';
 
 export default {
   test: element => {
