+++ src/shared/processors/low/removeArticles.js
