+++ src/shared/processors/low/removeDoctypeHeadTags.js
