+++ src/shared/processors/low/removeHidden.js
