+++ src/shared/processors/low/removeScript.js
