+++ src/shared/processors/low/removeTagStyle.js
