+++ src/shared/processors/low/removeWpEmbeddedContent.js
