+++ src/shared/processors/low/table.js
@@ -1,4 +1,4 @@
-import Table from '../components/Table';
+import Table from '../../components/Table';
 
 export default {
   test: ({ component }) => component === 'table',
