+++ src/shared/processors/medium/anchor.js
