+++ src/shared/processors/medium/myrGallery.js
