+++ src/shared/processors/medium/myrTiledGallery.js
