+++ src/shared/processors/medium/removeAmpColWidth.js
