+++ src/shared/processors/medium/removeAmpIds.js
