+++ src/shared/processors/medium/removeAmpListTypes.js
