+++ src/shared/processors/medium/removeInlineStyle.js
