+++ src/shared/processors/medium/tzGallery.js
